Refactor PlayTone to receive Note type
Changed the function PlayTone in Buzzer.go to receive a Note type argument instead of a float64 for note and duration. This makes the code cleaner and easier to read. Moreover, it emphasizes the relationship between the note (pitch) and its duration. Made corresponding updates in PlaySong function to maintain consistency.