Update Go dependencies and protocols for LED control
Updated the code to include additional Go dependencies and protocols in both the go.sum and go.mod files. This was done to facilitate the control flow in the LED.go script, allowing for LED routines to be stopped in the middle of execution through 'finish' and 'done' channels. This update also removed a TODO comment to tidy the code. These changes ensure improved responsiveness and user control within the system.