Add cross-compiling support to Makefile
Introduced a new Makefile to enable cross-compiling for the project. This allows building the application for different architectures (here specified as linux/arm64) directly from the host machine. Cross-compiling has been implemented using Docker, with a dedicated Docker command defined for this purpose. With these changes, the application can now be built to run on a Raspberry Pi machine.