Update fields to public and add streaming in rpc
In this update, we have revised the private fields in LED, Buzzer, and Ultrasonic scripts to be public fields. This is done to standardize the system and avoid potential access issues in the future.

Further, we also added streaming functionality to the Camera, ADC, and Ultrasonic scripts. This results in an improved and real-time experience for users.

Lastly, we added a new rpc file for the Camera. This will allow for integration of camera features within the application.

These changes will contribute to the effectiveness and usability of the system.