Remove .idea/watcherTasks.xml from repository
The .idea/watcherTasks.xml, a configuration file for a specific IDE, was removed from the repository. This file is specific to an individual developer's setups and isn't necessary for the project as a whole. Including it could lead to conflicts when other contributors use a different setup or IDE. To prevent this, it's been deleted. It is generally best practice to not include IDE-specific files in the repository.