Change car LED creation function to use pointer
The function 'CreateRGBStrip' in the 'LED.go' file of the 'Car' package has been renamed to 'NewRGBStrip', and now returns a pointer to an 'RGBStrip' instance, instead of directly returning an 'RGBStrip'. This change was implemented to follow Go's conventional practice of prefixing factory function names with 'New' and returning pointers from them. This can potentially help to optimize memory consumption, as the struct can now be modified in-place, without being copied each time.